MIT LICENSE
The MIT License (Massachusetts Institute of Technology License) is one of the most commonly used open-source software licenses. It is highly permissive, meaning that it allows anyone to freely use, modify, and distribute your project with minimal restrictions. The main requirement is that the original copyright notice and license must be included in any copies or substantial portions of the software.